requirements.txt
fastapi
uvicorn
requests
python-dotenv
tavily-python
streamlit